Move register method from AuthService to UserService
